Some problems with: view_folders_fun,def view_file_fun
